Improve performance by getting details of project by id.
Previously, we used the get_project_by_name() function without passing
a team identifier meaning it pulled the entire list of projects and
searched for a match. Now, when retrieving the scan data, we also
retrieve the project ID so that we can use it to retrieve the details
of the specific project.