The feature flags API needs the tenant_id as a filter
The feature flags API requires that the caller provide a "filter"
query parameter whose value is the tenant id. As we do not have the
tenant id in the configuration be default, we use the Keycloak API to
retrieve it the first time one of the feature flags endpoints is used.